Outgoing transfers controller can share functionality
This sets up the OutgoingTransfersController to share functionality with
the IncomingTransfersController, which will be done in the next commit. I've
also renamed the OutgoingTransferPresenter to be a TransferPresenter, so
it can be used for both types of Transfer
