+++ app/controllers/staff/activity_financials_controller.rb
@@ -16,6 +16,6 @@ class Staff::ActivityFinancialsController < Staff::BaseController
     @budget_presenters = @budgets.includes(:parent_activity, :providing_organisation).map { |budget| BudgetPresenter.new(budget) }
     @forecast_presenters = @forecasts.map { |forecast| ForecastPresenter.new(forecast) }
     @implementing_organisation_presenters = activity.implementing_organisations.map { |implementing_organisation| ImplementingOrganisationPresenter.new(implementing_organisation) }
-    @transfers = policy_scope(activity.source_transfers).map { |transfer| OutgoingTransferPresenter.new(transfer) }
+    @transfers = policy_scope(activity.source_transfers).map { |transfer| TransferPresenter.new(transfer) }
   end
 end
