+++ app/presenters/transfer_presenter.rb
@@ -1,4 +1,4 @@
-class OutgoingTransferPresenter < SimpleDelegator
+class TransferPresenter < SimpleDelegator
   def value
     return if super.blank?
     ActionController::Base.helpers.number_to_currency(super, unit: "£")
