+++ spec/presenters/transfer_presenter_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe OutgoingTransferPresenter do
+RSpec.describe TransferPresenter do
   let(:transfer) { build_stubbed(:outgoing_transfer, value: "110.01", financial_year: 2020, financial_quarter: 1) }
 
   describe "#financial_quarter_and_year" do
